make only code 2 screens
