+++ lib/home/business_logic/Home/qr_state.dart
@@ -1,20 +1,20 @@
-part of 'qr_cubit.dart';
-
-@immutable
-abstract class QrState {}
-
-class QrInitial extends QrState {}
-//QrLoading
-class QrLoading extends QrState {}
-//qrCodeScannedSuccessfully
-class QrCodeScannedSuccessfully extends QrState {
-}
-//QrError
-class QrError extends QrState {
-  final String message;
-
-  QrError(this.message);
-}
-//qrCodeScannedSuccessfully
-class QrCodeScannedSuccessfullyWithNoInternet extends QrState {
-}
+// part of 'qr_cubit.dart';
+//
+// @immutable
+// abstract class QrState {}
+//
+// class QrInitial extends QrState {}
+// //QrLoading
+// class QrLoading extends QrState {}
+// //qrCodeScannedSuccessfully
+// class QrCodeScannedSuccessfully extends QrState {
+// }
+// //QrError
+// class QrError extends QrState {
+//   final String message;
+//
+//   QrError(this.message);
+// }
+// //qrCodeScannedSuccessfully
+// class QrCodeScannedSuccessfullyWithNoInternet extends QrState {
+// }
