+++ lib/registeration/presenation/SignUpScreen.dart
@@ -8,11 +8,11 @@
 // import 'package:swimmer_app/registeration/business_logic/auth_cubit/sign_up_state.dart';
 // import 'package:swimmer_app/registeration/presenation/otp.dart';
 // import 'package:swimmer_app/registeration/presenation/widget/component.dart';
-// import 'package:swimmer_app/registeration/presenation/widget/widget.dart';
+// import 'package:swimmer_app/registeration/presenation/widget/bottom_nav_bar.dart';
 //
 // import '../../core/constants/font_manager.dart';
 // import '../../core/constants/styles_manager.dart';
-// import '../../home/presenation/widget/widget.dart';
+// import '../../home/presenation/widget/bottom_nav_bar.dart';
 //
 // class SignUpScreen extends StatelessWidget {
 //
