+++ lib/routiong.dart
@@ -2,16 +2,14 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:swimmer_app/home/business_logic/Home/dash_board_cubit.dart';
 import 'package:swimmer_app/home/presenation/home_lauout.dart';
 import 'package:swimmer_app/registeration/business_logic/auth_cubit/login_cubit.dart';
-import 'package:swimmer_app/registeration/presenation/SignUpScreen.dart';
 import 'package:swimmer_app/registeration/presenation/welcome_screen.dart';
 import 'package:swimmer_app/registeration/presenation/login_screen.dart';
-import 'package:swimmer_app/registeration/presenation/otp.dart';
 
 
 import 'core/constants/routes_manager.dart';
-import 'core/constants/strings.dart';
 
 class RouteGenerator {
   static Route<dynamic>? generateRoute(RouteSettings settings) {
@@ -27,19 +25,12 @@ class RouteGenerator {
               child: SignInScreen(),
             ));
       case AppRoutes.home:
-        return MaterialPageRoute(builder: (_) => HomeLayout());
-    //EditProfile
-      case AppRoutes.editProfile:
-        return MaterialPageRoute(builder: (_) => EditProfile());
-      // default:
-      //   return _errorRoute();
+        return MaterialPageRoute(builder: (_) =>
+            BlocProvider(
+              create: (context) => DashBoardCubit()..getUserData(),
+              child: HomeLayout(),
+            ));
     }
   }
-  static Route<dynamic> _errorRoute() {
-    // Exit the app if there is no route defined for the given route settings
-    return MaterialPageRoute(builder: (_) {
-      WidgetsBinding.instance.addPostFrameCallback((_) => exit(0));
-      return Container();
-    });
-  }
+
 }
